requirements.txt
streamlit
numpy
faiss-cpu
groq
python-dotenv
